Přihlášení

+++ src/main/java/org/example/domacirozpocet2/repository/RoleRepository.java
@@ -1,10 +1,9 @@
 package org.example.domacirozpocet2.repository;
 
 import org.example.domacirozpocet2.entity.Role;
-import org.example.domacirozpocet2.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository <Role, Integer> {
+public interface RoleRepository extends JpaRepository <Role, Long> {
     Role findByName(String name);
 
 }
